requirements.txt
streamlit==1.27.0
pandas==2.0.3
plotly==5.17.0
folium==0.14.0
streamlit-folium==0.13.0
google-generativeai==0.3.1
python-dotenv==1.0.0
nest-asyncio==1.5.6
pickle-mixin==1.0.2
requests==2.31.0
